producer_app/producer_app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asttokens==3.0.0
    # via stack-data
black[jupyter]==25.1.0
    # via -r requirements.in
click==8.1.8
    # via black
decorator==5.2.1
    # via ipython
exceptiongroup==1.2.2
    # via ipython
executing==2.2.0
    # via stack-data
ipython==8.32.0
    # via black
jedi==0.19.2
    # via ipython
lz4==4.4.3
    # via mcap
matplotlib-inline==0.1.7
    # via ipython
mcap==1.2.2
    # via
    #   -r requirements.in
    #   mcap-protobuf-support
mcap-protobuf-support==0.5.3
    # via -r requirements.in
multidict==6.1.0
    # via pulsar
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via black
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
platformdirs==4.3.6
    # via black
prompt-toolkit==3.0.50
    # via ipython
protobuf==5.29.3
    # via mcap-protobuf-support
ptyprocess==0.7.0
    # via pexpect
pulsar==2.0.2
    # via -r requirements.in
pure-eval==0.2.3
    # via stack-data
pygments==2.19.1
    # via ipython
stack-data==0.6.3
    # via ipython
tokenize-rt==6.1.0
    # via black
tomli==2.2.1
    # via black
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   black
    #   ipython
    #   multidict
wcwidth==0.2.13
    # via prompt-toolkit
zstandard==0.23.0
    # via mcap
